requirements.txt

Pillow (3.4.2)
MySQL-python (1.2.4b4)
Flask (0.11.1)
Flask-Login (0.3.2)
Flask-Mail (0.9.1)
Flask-Migrate (2.0.1)
Flask-Principal (0.4.0)
Flask-Script (2.0.5)
Flask-Security (1.7.5)
Flask-SQLAlchemy (2.1)
Flask-WTF (0.13.1)
python-editor (1.0.3)
python-memcached (1.58)
qiniu (7.0.9.1)
WTForms (2.1)
shortuuid (0.4.3)